Refactor: split background worker by concern (tracker / router / lifecycle / badge)
`background.ts` had grown to 672 lines mixing five concerns: per-tab
in-memory state, badge/icon rendering, state mutators, the typed message
router, and the chrome.tabs / storage lifecycle wiring. Split into focused
modules under `lib/background/`, leaving the entry point as 43 lines of
pure wiring.

- `badge.ts` — toolbar badge/icon painting (chrome.action side effects only)
- `tab-tracker.ts` — `createTabTracker()` owns all per-tab state + the badge
  refresh that reads across it
- `message-handlers.ts` — `registerBackgroundMethods(tracker)`, the typed
  content/popup→background router
- `lifecycle.ts` — `startBackgroundLifecycle(tracker)`: tab listeners,
  storage seeds/subscriptions, the recovery-pause session bridge

Behavior-preserving: the enforcement subscription and its startup seed are
unified onto `setEnforcementEnabled` (the clear-stale-maps branch is a no-op
on empty maps at seed time). Session-store writes and content broadcasts stay
in the lifecycle layer; the tracker owns only in-memory cache.

Background-purity check still passes (new modules import only `RuleId` from
`rules/rule-metadata`). A side benefit: the badge/tracker helpers were locked
in the untestable entry bundle and are now unit-testable.
